Do not include whitespace between .add-on and input
If you do, there is a whitespace gap between the elements as they are
being displayed as "inline-block". Tricks with the word-spacing CSS property
are inconsistent and best avoided, especially when we can just render HTML
that works cross-browser without the user having to write it any different.

Prior art: the Twitter Bootstrap docs use this method.